Autoloader: load plugins on muplugins_loaded
Directly calling `loadPlugins` wasn't correct since all the autoloaded
plugins would be loaded when the Autoloader was loaded. If another
muplugin happened to be loaded before it, it wouldn't have access to the
rest of the plugins.

Using the `muplugins_loaded` action ensures consistency. All the
autoloaded plugins will be guaranteed to be loaded at the same part of
the loading process now and have access to all regular muplugins.