Add missing composer step
This is strange because the composer install happens as part of the docker
build. We use a named volume on top of /vendor, which would be empty. I thought
that the volume would just be layered on top, so the docker container would still
have content in it.

I wonder if this behavior is different on linux vs macos.